# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Atlas Treasury API
  version: 0.2.0
  description: Transfers, reserves, holds, reconciliation, and settlement operations
  contact:
    name: AtlasPayments Team
    url: https://github.com/shivemind/atlaspayments
servers:
  - url: http://localhost:3000
    description: Local development server. Override via Postman {{baseUrl}} environment variable.
...
paths:
  /api/v1/transfers:
    get:
      summary: List transfers
      description: Returns a paginated list of transfers for the authenticated merchant.
      x-platform:
        kind: report
        entity: Transfer
        affectsLedger: false
      operationId: listTransfers
      tags:
        - Transfers
      security:
        - MerchantKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
      responses:
        "200":
          description: Paginated list of transfers.
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - pagination
                properties:
                  data:
                    type: array
                    items:
                      type: object
                  pagination:
                    $ref: "#/components/schemas/PaginationMeta"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: Create a transfer
      description: Creates a new transfer between accounts for the authenticated merchant.
      x-platform:
        kind: transition
        entity: Transfer
        affectsLedger: true
        stateChanges:
          - created
      operationId: createTransfer
      tags:
        - Transfers
      security:
        - MerchantKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - amount
                - currency
              properties:
                amount:
                  type: integer
                  minimum: 1
                currency:
                  type: string
                  minLength: 3
                  maxLength: 3
                description:
                  type: string
                metadata:
                  type: object
                  additionalProperties: true
      responses:
        "201":
          description: Transfer created.
          content:
            application/json:
              schema:
                type: object
                required:
                  - transfer
                properties:
                  transfer:
                    type: object
        "400":
          description: Invalid request payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Idempotency key conflict.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/transfers/{id}:
    get:
      summary: Get transfer by ID
      description: Retrieves a single transfer by its ID.
      x-platform:
        kind: report
        entity: Transfer
        affectsLedger: false
      operationId: getTransferById
      tags:
        - Transfers
      security:
        - MerchantKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: Transfer details.
          content:
            application/json:
              schema:
                type: object
                required:
                  - transfer
                properties:
                  transfer:
                    type: object
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Transfer not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      summary: Update transfer metadata
      description: Updates mutable fields on an existing transfer.
      x-platform:
        kind: transition
        entity: Transfer
        affectsLedger: false
        stateChanges:
          - metadata_updated
      operationId: updateTransfer
      tags:
        - Transfers
      security:
        - MerchantKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                description:
                  type: string
                metadata:
                  type: object
                  additionalProperties: true
      responses:
        "200":
          description: Transfer updated.
          content:
            application/json:
              schema:
                type: object
                required:
                  - transfer
                properties:
                  transfer:
                    type: object
        "400":
          description: Invalid request payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Transfer not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/v1/reserves:
    get:
      summary: List reserves
      description: Returns a paginated list of reserves for the authenticated merchant.
      x-platform:
        kind: report
        entity: Reserve
        affectsLedger: false
      operationId: listReserves
      tags:
        - Reserves
      security:
        - MerchantKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
      responses:
        "200":
          description: Paginated list of reserves.
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - pagination
                properties:
                  data:
                    type: array
                    items:
                      type: object
                  pagination:
                    $ref: "#/components/schemas/PaginationMeta"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: Create a reserve
      description: Creates a reserve to hold funds for the authenticated merchant.
      x-platform:
        kind: transition
        entity: Reserve
        affectsLedger: true
        stateChanges:
          - created
      operationId: createReserve
      tags:
        - Reserves
      security:
        - MerchantKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - amount
                - currency
              properties:
                amount:
                  type: integer
                  minimum: 1
                currency:
                  type: string
                  minLength: 3
                  maxLength: 3
                reason:
                  type: string
                metadata:
                  type: object
                  additionalProperties: true
      responses:
        "201":
          description: Reserve created.
          content:
            application/json:
              schema:
                type: object
                required:
                  - reserve
                properties:
                  reserve:
                    type: object
        "400":
          description: Invalid request payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Idempotency key conflict.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    MerchantKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: |
        Bearer API key for merchant-scoped operations. Pass as `Authorization: Bearer {{merchantApiKey}}`.
    PlatformAdminKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: |
        Bearer API key for platform admin operations. Pass as `Authorization: Bearer {{platformAdminKey}}`.
  schemas:
    ErrorResponse:
      description: Standard error response envelope returned for 4xx/5xx responses.
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Machine-readable error code.
            message:
              type: string
              description: Human-readable error description.
    PaginationMeta:
      description: Pagination metadata included in list responses.
      type: object
      required:
        - page
        - pageSize
        - total
        - hasMore
      properties:
        page:
          type: integer
          description: Current page number (1-indexed).
        pageSize:
          type: integer
          description: Number of results per page.
        total:
          type: integer
          description: Total number of matching records.
        hasMore:
          type: boolean
          description: Whether additional pages exist.
...
  parameters:
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      description: |
        Client-generated unique key for safe request retries. Identical key + payload replays the original response.
      schema:
        type: string
        example: idem-key-a1b2c3d4
    PageParam:
      in: query
      name: page
      description: Page number (1-indexed).
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      in: query
      name: pageSize
      description: Number of results per page (max 100).
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
    PathId:
      in: path
      name: id
      required: true
      description: Resource identifier.
      schema:
        type: string
...